Suggestion for folder- and file-structure for ts
