+++ src/ts/config.ts
@@ -1,0 +1,2 @@
+export const BASE_URL = "http://localhost:3000/api";
+export const API_KEY = "group3api";
